keep push notification alive
currenttly all browser requrire a `push` produces a notification, if
not, push subscription may be revoked.

see:
https://github.com/firebase/firebase-js-sdk/issues/8010#issuecomment-2119275829

in this case, sygnal, the push gateway will send unread count, and
create a `push` event, when messages are read on other clients.
Therefore we have no choice but to stop sygnal from pushing data that
won't be presented as notifications

see:
https://github.com/element-hq/sygnal/blob/da48ba7457409391c1d12140ce2ea3d49d377ff3/sygnal/webpushpushkin.py#L142-L145